requirements.txt
# Python 3.14 Compatible Requirements

# FastAPI without Pydantic v2 dependency
fastapi==0.109.2
starlette==0.36.3
python-multipart==0.0.9

# Server
uvicorn==0.27.1
websockets==12.0

# HTTP Clients
requests==2.31.0
httpx==0.26.0

# Hugging Face
huggingface-hub==0.20.3

# Audio
numpy==1.26.4
soundfile==0.12.1

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1

# Logging and Monitoring
loguru==0.7.2
psutil==5.9.8  # System resource monitoring
aiohttp==3.9.1  # Async HTTP for webhooks/alerts

# Type checking
typing-extensions==4.15.0

# Asterisk ARI Integration
ari==0.1.3
aiortc==1.6.0  # WebRTC/RTP handling for media streams
